requirements.txt
pyyaml>=6.0
psutil>=5.9.0
pynput>=1.7.6
pyautogui>=0.9.54
click>=8.0.0
colorama>=0.4.6
pytest>=7.4.0
pytest-cov>=4.1.0

# Platform-specific dependencies
pyobjc-framework-Cocoa>=9.0; sys_platform == "darwin"
pyobjc-framework-Quartz>=9.0; sys_platform == "darwin"
python-xlib>=0.33; sys_platform == "linux"
pywin32>=306; sys_platform == "win32"
